+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add version number to tools, such as idm -v, derived from xenon/config.h
 - Upgrade 36.331 to r15
 - Build: CI testing with Travis now
 - Build: Using c++17 and no longer require boost for tests
@@ -17,7 +18,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Build: Use meson instead of cmake now
 - Doc: Updated README
 - Doc: Generated XDDL documentation updated